+++ server/src/macroLanguageService/macroLanguageService.ts
@@ -9,11 +9,12 @@ import { MacroHover } from './services/macroHover';
 import { MacroNavigation as MacroNavigation } from './services/macroNavigation';
 import { MacroValidation } from './services/macroValidation';
 import { MacroCompletion } from './services/macroCompletions';
+import { MacroCommand } from './services/macroCommands';
 
 import {
 	LanguageSettings, LanguageServiceOptions, DocumentContext, DocumentLink,
 	SymbolInformation, Diagnostic, Position, Hover, Location, 
-	TextDocument, CompletionList, CodeLens
+	TextDocument, CompletionList, CodeLens, TextDocumentEdit
 } from './macroLanguageTypes';
 
 
@@ -21,19 +22,21 @@ export type Macrofile = {};
 export * from './macroLanguageTypes';
 
 export interface LanguageService {
-	doValidation(document: TextDocument, file: Macrofile, documentSettings?: LanguageSettings): Diagnostic[];
+	doValidation(document: TextDocument, file: Macrofile, documentSettings: LanguageSettings): Diagnostic[];
 	parseMacroFile(document: TextDocument): Macrofile;
 	doHover(document: TextDocument, position: Position, macroFile: Macrofile):Hover | null;
-	doComplete(document: TextDocument, position: Position, stylesheet: Macrofile): CompletionList;
+	doComplete(document: TextDocument, position: Position, stylesheet: Macrofile, documentSettings: LanguageSettings): CompletionList;
 	findDefinition(document: TextDocument, position: Position, macroFile: Macrofile): Location | null;
 	findReferences(document: TextDocument, position: Position, macroFile: Macrofile): Location[];
 	findImplementations(document: TextDocument, position: Position, macroFile: Macrofile): Location[];
 	findDocumentLinks(document: TextDocument, macrofile: Macrofile, documentContext: DocumentContext): DocumentLink[];
 	findDocumentSymbols(document: TextDocument, macrofile: Macrofile): SymbolInformation[];
 	findCodeLenses(document: TextDocument, macrofile: Macrofile): CodeLens[];
+	doRefactorSequences(document: TextDocument, position: Position, macrofile: Macrofile, documentSettings: LanguageSettings) : TextDocumentEdit | null;
+	doCreateSequences(document: TextDocument, position: Position, macrofile: Macrofile, documentSettings: LanguageSettings) : TextDocumentEdit | null;
 }
 
-function createFacade(parser: Parser, hover: MacroHover, completion: MacroCompletion, navigation: MacroNavigation, validation: MacroValidation): LanguageService {
+function createFacade(parser: Parser, hover: MacroHover, completion: MacroCompletion, navigation: MacroNavigation, validation: MacroValidation, command: MacroCommand): LanguageService {
 	return {
 		doValidation: validation.doValidation.bind(validation),
 		parseMacroFile: parser.parseMacroFile.bind(parser),
@@ -44,7 +47,9 @@ function createFacade(parser: Parser, hover: MacroHover, completion: MacroComple
 		findImplementations: navigation.findImplementations.bind(navigation),
 		findDocumentLinks: navigation.findDocumentLinks.bind(navigation),
 		findDocumentSymbols: navigation.findDocumentSymbols.bind(navigation),
-		findCodeLenses: navigation.findCodeLenses.bind(navigation)
+		findCodeLenses: navigation.findCodeLenses.bind(navigation),
+		doRefactorSequences: command.doRefactorSequences.bind(command),
+		doCreateSequences: command.doCreateSequences.bind(command)
 	};
 }
 
@@ -55,6 +60,7 @@ export function getMacroLanguageService(options: LanguageServiceOptions): Langua
 		new MacroHover(options && options.fileProvider),
 		new MacroCompletion(options && options.fileProvider),
 		new MacroNavigation(options && options.fileProvider),
-		new MacroValidation(options && options.fileProvider)
+		new MacroValidation(options && options.fileProvider),
+		new MacroCommand(options && options.fileProvider),
 	);
 }