+++ server/src/macroLanguageService/services/macroCompletions.ts
@@ -13,7 +13,10 @@ import {
 	CompletionItem, 
 	CompletionItemKind, 
 	Range, 
-	MacroFileProvider
+	MacroFileProvider,
+	InsertTextFormat,
+	TextEdit,
+	LanguageSettings
 } from '../macroLanguageTypes';
 import { getComment } from '../parser/macroScanner';
 
@@ -85,6 +88,7 @@ export class MacroCompletion {
 	symbolContext!: Symbols;
 	defaultReplaceRange!: Range;
 	nodePath!: nodes.Node[];
+	inc : number = 10;
 
 	constructor(private fileProvider: MacroFileProvider) {}
 
@@ -95,7 +99,7 @@ export class MacroCompletion {
 			this.symbolContext = new Symbols(this.macroFile, this.textDocument.uri);
 
 			// Included symbols 
-			const uris = this.getIncludeUris(this.macroFile, this.fileProvider);
+			const uris = this.getIncludeUris(this.macroFile);
 			const types = this.fileProvider?.getAll({uris:uris});
 			for (const type of types) {
 				const symbols = new Symbols(<nodes.Node>type.macrofile, type.document.uri);
@@ -105,17 +109,18 @@ export class MacroCompletion {
 		return this.symbolContext;
 	}
 
-	private getIncludeUris(macroFile: nodes.MacroFile, fileProvider:MacroFileProvider) : string[] {
+	private getIncludeUris(macroFile: nodes.MacroFile) : string[] {
 		return <string[]>macroFile.getData(nodes.Data.Includes);
 	}
 
-	public doComplete(document: TextDocument, position: Position, macroFile: nodes.MacroFile): CompletionList {
+	public doComplete(document: TextDocument, position: Position, macroFile: nodes.MacroFile, settings: LanguageSettings): CompletionList {
 		this.offset = document.offsetAt(position);
 		this.position = position;
 		this.currentWord = this.getCurrentWord(document, this.offset);
 		this.defaultReplaceRange = Range.create(Position.create(this.position.line, this.position.character - this.currentWord.length), this.position);
 		this.textDocument = document;
 		this.macroFile = macroFile;
+		this.inc	= settings.sequence?.increment ? Number(settings.sequence?.increment) : 10;
 
 		try {
 			const result: CompletionList = { isIncomplete: false, items: [] };
@@ -128,11 +133,13 @@ export class MacroCompletion {
 					this.getSymbolProposals(result, nodes.ReferenceType.Variable);
 					this.getSymbolProposals(result, nodes.ReferenceType.Label);
 					this.getKeyWordProposals(result);
+					this.getSequenceNumberSnipped(node, result);
 				}
 				else if (node.type === nodes.NodeType.If || node.type === nodes.NodeType.While) {
 					this.getSymbolProposals(result, nodes.ReferenceType.Variable, macroStatementTypes);
 					this.getSymbolProposals(result, nodes.ReferenceType.Label, labelTypes);
 					this.getKeyWordProposals(result);
+					this.getSequenceNumberSnipped(node, result);
 				}
 				else if (node.type === nodes.NodeType.Condition || node.type === nodes.NodeType.BinaryExpression) {
 					this.getSymbolProposals(result, nodes.ReferenceType.Variable, macroStatementTypes);
@@ -275,6 +282,37 @@ export class MacroCompletion {
 		return result;
 	}
 
+	private getSequenceNumberSnipped(node: nodes.Node, result: CompletionList): CompletionList {
+
+		const func = <nodes.Function>node.findAParent(nodes.NodeType.Function);
+		let seq = 0;
+		if (func) {
+			func.accept(candidate => {
+				if (candidate.type === nodes.NodeType.SequenceNumber) {
+					const nnode = (<nodes.SequenceNumber>candidate).getNumber();
+					if (nnode) {
+						const number = nnode.getText().toLocaleLowerCase().split('n').pop();
+						seq = Math.max(seq, Number(number) + this.inc)
+						return false;
+					}
+				}
+				return true;
+			});
+		}
+
+		const range = this.getCompletionRange(null);
+		const item: CompletionItem = {
+			label: 'N-Number',
+			documentation: '',
+			textEdit: TextEdit.replace(range, 'N${1:'+`${seq}`+'} $0'),
+			insertTextFormat: InsertTextFormat.Snippet,
+			kind: CompletionItemKind.Snippet,
+			sortText: ' '
+		};
+		result.items.push(item);
+		return result;
+	}
+
 	private getCompletionsForMacroFile(result: CompletionList): CompletionList {
 		const node = this.macroFile.findFirstChildBeforeOffset(this.offset);
 		for (const key of keys) {
@@ -290,6 +328,17 @@ export class MacroCompletion {
 			i--;
 		}
 		return text.substring(i + 1, offset);
+	}
+
+	private getCompletionRange(existingNode: nodes.Node | null) : Range {
+		if (existingNode && existingNode.offset <= this.offset && this.offset <= existingNode.end) {
+			const end = existingNode.end !== -1 ? this.textDocument.positionAt(existingNode.end) : this.position;
+			const start = this.textDocument.positionAt(existingNode.offset);
+			if (start.line === end.line) {
+				return Range.create(start, end); // multi line edits are not allowed
+			}
+		}
+		return this.defaultReplaceRange;
 	}
 }